CODE REFACTOR
So remember to use the strip() and lower() methods; strip() removes whitespace!
I also changed it to check if yes is in the string, not if the string is equal to yes; In the future i might build a simple AI lib for this, lol. not real ai, just scanning a huge list.